test: migrate process module specs to spec-main
